+++ jaxb-ri/xjc/src/main/java/com/sun/tools/xjc/reader/xmlschema/ct/Messages.java
@@ -46,16 +46,16 @@ import java.util.ResourceBundle;
 /**
  * Message resources
  */
-enum Messages {
-    ERR_NO_FURTHER_EXTENSION;
+public enum Messages {
+	ERR_NO_FURTHER_EXTENSION;
 
-    private static final ResourceBundle rb = ResourceBundle.getBundle(Messages.class.getPackage().getName() + ".MessageBundle");
+	private static final ResourceBundle rb = ResourceBundle.getBundle(Messages.class.getPackage().getName() + ".MessageBundle");
 
-    public String toString() {
-        return format();
-    }
+	public String toString() {
+		return format();
+	}
 
-    public String format( Object... args ) {
-        return MessageFormat.format( rb.getString(name()), args );
-    }
+	public String format(Object... args) {
+		return MessageFormat.format(rb.getString(name()), args);
+	}
 }
