+++ jaxb-ri/xjc/src/main/java/com/sun/tools/xjc/reader/xmlschema/ct/clFork/MyMixedExtendedComplexTypeBuilder.java
@@ -38,35 +38,65 @@
  * holder.
  */
 
-package com.sun.tools.xjc.reader.xmlschema.ct;
+package com.sun.tools.xjc.reader.xmlschema.ct.clFork;
 
 import com.sun.tools.xjc.model.CClass;
-import com.sun.tools.xjc.model.CClassInfo;
+import com.sun.tools.xjc.model.CPropertyInfo;
+import com.sun.tools.xjc.reader.RawTypeSet;
+import com.sun.tools.xjc.reader.xmlschema.RawTypeSetBuilder;
+import com.sun.tools.xjc.reader.xmlschema.bindinfo.BIProperty;
+import com.sun.tools.xjc.reader.xmlschema.ct.AbstractExtendedComplexTypeBuilder;
+import com.sun.tools.xjc.reader.xmlschema.ct.ComplexTypeBindingMode;
+import com.sun.tools.xjc.reader.xmlschema.ct.Messages;
 import com.sun.xml.xsom.XSComplexType;
 import com.sun.xml.xsom.XSType;
 
 /**
- * Binds a complex type derived from another complex type by restriction.
- *
+ * Handles the Mixed extensions....better?
  */
-final class MyRestrictedComplexTypeBuilder extends CTBuilder {
+public final class MyMixedExtendedComplexTypeBuilder extends AbstractExtendedComplexTypeBuilder {
 
 	public boolean isApplicable(XSComplexType ct) {
-		XSType baseType = ct.getBaseType();
-		return baseType != schemas.getAnyType() && baseType.isComplexType() && ct.getDerivationMethod() == XSType.RESTRICTION;
+
+		// TODO: is this necessary?
+		if (!bgmBuilder.isGenerateMixedExtensions())
+			return false;
+
+		XSType bt = ct.getBaseType();
+		if (bt.isComplexType() && bt.asComplexType().isMixed() && ct.getDerivationMethod() == XSType.EXTENSION && ct.isMixed()
+				&& ct.getContentType().asParticle() != null && ct.getExplicitContent().asEmpty() == null) {
+			return true;
+		}
+
+		return false;
 	}
 
 	public void build(XSComplexType ct) {
-
-		new FreshComplexTypeBuilder().build(ct);
 		XSComplexType baseType = ct.getBaseType().asComplexType();
 
 		// build the base class
 		CClass baseClass = selector.bindToType(baseType, ct, true);
 		assert baseClass != null; // global complex type must map to a class
 
-		CClassInfo currentBean = selector.getCurrentBean();
-		BaseClassManager m = BaseClassManager.getInstance();
-		m.createRestrictingClass(currentBean, baseClass);
+		if (!checkIfExtensionSafe(baseType, ct)) {
+			// error. We can't handle any further extension
+			errorReceiver.error(ct.getLocator(), Messages.ERR_NO_FURTHER_EXTENSION.format(baseType.getName(), ct.getName()));
+			return;
+		}
+
+		selector.getCurrentBean().setBaseClass(baseClass);
+		builder.recordBindingMode(ct, ComplexTypeBindingMode.FALLBACK_EXTENSION);
+
+		BIProperty prop = BIProperty.getCustomization(ct);
+		CPropertyInfo p;
+
+		RawTypeSet ts = RawTypeSetBuilder.build(ct.getContentType().asParticle(), false);
+		p = prop.createContentExtendedMixedReferenceProperty(ct.getName() + "_Mixed", ct, ts);
+
+		selector.getCurrentBean().addProperty(p);
+
+		// adds attributes and we are through.
+		green.attContainer(ct);
 	}
+
 }
